Update recovery.fstab format.
In Android 4.3 the various fstab files were unified in a common file.
Although CWM still uses the "/etc/recovery.fstab" file, the format of
entries in that file is now the same as the format of entries in the
common fstab file:
<source> <mount point> <type> <mount options> <fs_mgr flags>

Every field is mandatory; if no special mount options or fs_mgr flags
are needed, "defaults" must be used.

The mount options for internal partitions are based on those used in the
official Fairphone image. However, they are not exactly the same, as
"noauto_da_alloc" seemed less safe than the default "auto_da_alloc"
(although in the case of the recovery it probably would not matter), and
"commit=1" seemed unnecessary for the "/system" partition (or any other
partition in the recovery, as in the event of a power loss while
restoring a backup the backup could be just restored again when the
device is switched on).

The fs_mgr flags for internal partitions just ensure that the devices
have appeared before trying to mount them and that their filesystem is
fine, and marks them as non removable.

The mount options for Vold managed partitions are set to "defaults" as
Vold ignores them, and the fs_mgr flags mark them as "managed by Vold".
The file system for external partitions is set the "auto" to let Vold
find the right one.

Finally, the source for Vold managed partitions is a device from
"sysfs". Vold listens to hotplug messages sent by the kernel and
compares the start of the "DEVPATH" parameter in the message with the
sources to find the associated partition, if any.