Add sharing for both the internal and external SD card.
With the default LUN pattern used by Vold only a single partition could
be shared (the internal SD card, as it is the first one in the fstab).
However, the Fairphone kernel provides another LUN file that could be
used to share also the external SD card. In CyanogenMod (although not in
AOSP), Vold supports defining a custom LUN pattern besides the default
one to share more volumes, so TARGET_USE_CUSTOM_LUN_FILE_PATH was set to
the appropriate value.