Fix (workaround) framebuffer pixel format.
The free gralloc is expected to be used with libagl, the software OpenGL
ES library from Android. For 32-bit pixel formats libagl only supports
RGB with offsets R=0, G=8 and B=16, but the MediaTek framebuffer kernel
driver, due to a bug, only supports RGB with offsets R=16, G=8 and B=0.
However, as both support 16-bit RGB with offsets R=11, G=5 and B=0 the
framebuffer is forced to that pixel format. Otherwise, red and blue
channels would be swapped and therefore colors would be incorrect.

The real culprit here, though, is a bug in MediaTek framebuffer kernel
driver. Theoretically, the framebuffer kernel driver can be configured
to use a 32-bit pixel format with offsets R=0, G=8 and B=16. But, in
practice, the given offsets for 32-bit pixel formats are ignored and
they are always forced to their default value, R=16, G=8 and B=0. As the
bug is deeply rooted in the driver ("mtkfb_pan_display_impl" overwrites
the format set by "mtkfb_set_overlay_layer" in a shared variable between
them with a value that does not even have the right type) fixing it may
snowball and affect other components using the framebuffer. It seems
safer, at least for the time being, to just add a workaround for it in
this module instead of fixing it and maybe breaking other clients of the
driver. In fact, taking a look to the code in the official Android
kernel for MediaTek chipsets it seems that it was not fixed until the
"android-mtk-3.18" branch, not even for Marshmallow on sprout.

As an aside, it is worth noting that Android and the code of the
MediaTek framebuffer kernel driver use different names to refer to the
same pixel formats. For example, the pixel format with offsets R=0, G=8,
B=16 and A=24 in Android is RGBA, but in MediaTek framebuffer kernel
driver is ABGR. In the same way, the pixel format with offsets R=16,
G=8, B=0 and A=24 in Android is BGRA, while in MediaTek framebuffer
kernel driver is ARGB.