Fix main system when using software OpenGL ES.
When software OpenGL ES is used the SystemUI is restarted over and over.
The problem comes from ImageWallpaper, which crashes at start due to
requesting an OpenGL ES 2.0 context and not finding a suitable eglConfig
(as the software OpenGL ES only supports OpenGL ES 1.0). However,
ImageWallpaper only tries to use OpenGL ES 2.0 when
ActivityManager.isHighEndGfx() returns true; if not, it just happily
draws to a Canvas, and therefore works implicitly with just OpenGL ES
1.0.

This problem is addressed by the commit a5ed31ba3b in the
frameworks/base repository from Replicant, in which isHighEndGfx is
modified to return false when the property "ro.softwaregl" is set to
true. That commit also disables hardware rendering and uses a lighter
screen rotation animation, which do not fix specific problems, but it
should improve the performance of the system.

Unfortunately, those changes must be applied to the frameworks/base
repository; otherwise it is not possible to fix the problems in the
main system when using software OpenGL ES. Due to this, a special system
to automatically patch repositories from the full tree when needed
before building was introduced. It is an ugly hack, but it is a
necessary hack :(