Add missing TARGET_BOOTLOADER_BOARD_NAME build parameter.
The TARGET_BOOTLOADER_BOARD_NAME build parameter is used to define the
property "ro.product.board". This property is used by libhardware as one
of the variant for the filenames of modules, with higher priority than
"ro.board.platform".

Basically, it makes possible to define device specific modules in the
build system whose name do not clash with other possible modules in the
build system; if not defined, Fairphone 1 specific modules would have to
be called, for example, "gralloc.mt6589", which may clash with a generic
gralloc module for the MT6589 platform.