Add support for blink in lights module.
Blink is done using the "timer" trigger in the LED class of the kernel.
Unfortunately, it does not seem to be possible to synchronize several
blinking LEDs, so quick blink with combined colors does not look too
good.

Moreover, the lights module implementation is pretty dumb. It is mostly
what you set is what you get; it does not take into account the previous
state of the LEDs to try to deduce if the new color should be merged
with the previous one or override it, or to set just some components of
the color to blink and keep the other static (for example, if the red
LED was on and amber blink was requested). There is, however, one
exception: if there is a static color and then a blink color whose
components are off in the static color is requested the static color is
kept and the blink is done only in the components of the new color. This
makes possible, for example, to keep the red battery light on while
showing a blue blinking notification (on the other hand, this "feature"
is an unfortunate coupling to the behavior of upper layers... but for
the time being it seems like a good compromise).