Add USB configurations for mass_storage, mtp and ptp.
The idVendor for the USB configurations was set to 18D1, which is the
idVendor used by default in init.usb.rc for other functions (it is the
USB vendor ID of Google). The idProducts start at 4D00 because... well,
because it is similar to the 2D00 product id used for an Android-powered
device in accessory mode and it seemed to be an unused range.