Add optional support for proprietary features.
Unfortunately, this FOSS device tree does not provide support for all
the hardware features in the device, and support for some of them
requires the use of proprietary software provided in a separate vendor
tree. Now the system will use that proprietary software if the vendor
tree for the Fairphone 1 exists at the time of building but, if it does
not, it will just build and use only the FOSS device tree as before.