Update softwaregl patch to Android 4.3.
ScreenRotationAnimation.java is no longer patched, as the patched
version offers almost no performance benefit over the default animation
and it does not look as good.

The android.os.Build class is now patched to make possible for
applications to know whether the software OpenGL ES library is being
used or not (as SystemProperties is a hidden class). This will be
necessary, for example, for the Gallery to use OpenGL ES 1 instead of
OpenGL ES 2 in a future patch.