Fix Gallery2 when using software OpenGL ES.
The Gallery2 app supports OpenGL ES 1.0 and OpenGL ES 2.0, but it
assumes that if OpenGL ES 2.0 is required for the current system version
then the system has OpenGL ES 2.0 support. Unfortunately, this is not
the case when using the software OpenGL ES library, so the app must be
patched to also take into account whether software OpenGL ES is being
used or not to decide the OpenGL ES version of the EGL context to ask
for.