Adapt storage to changes in Android 4.4.
In 4.4, Vold no longer mounts its volumes directly in the mount point
defined in the fstab (Vold itself points out that those mount points
should be set to "auto"). Instead, volumes are mounted now in a two-step
process: first, the volume is directly mounted in the <volume label>
subdirectory inside the media directory, "/mnt/media_rw", which is a
directory that only the media_rw user has access to. Then, Vold starts
the fuse_<volume label> init service, which is the one that exposes the
volume in its final location and with the appropriate permissions.