Fix building an OTA package without a recovery image.
When building an OTA package, the build system in CyanogenMod 11.0
expects a recovery image to be built too. The "ota_from_target_files"
script uses the recovery image passed to it or builds one itself if none
is passed, but when a custom "mkbootimg" is needed to build the image,
using the passed image is the only valid approach, as the script has no
support for a custom "mkbootimg" and it always builds the images using
the default one.

Therefore, when BOARD_CUSTOM_BOOTIMG_MK is defined in order to use a
custom "mkbootimg" the recovery image is always passed to the
"ota_from_target_files" script. However, if TARGET_NO_RECOVERY is also
defined no recovery image is built, so it can not be passed to the
script and the build fails.

The recovery image is used in the "ota_from_target_files" script for the
"two-step" option and to patch the recovery partition based on the boot
partition in a running system. Neither the "two-step" OTA package nor
the recovery patching from boot are a must for the Fairphone 1, so the
build system can be modified to ignore the recovery image when building
an OTA package and TARGET_NO_RECVOERY is defined. The build system in
CyanogenMod 11.0 provides the TARGET_NO_SEPARATE_RECOVERY parameter
that, while semantically is not the same, has the desired effect of
mostly ignoring the recovery image, albeit not everywhere, so the build
system has to be modified anyway.

Unfortunately, those changes must be applied to the build repository;
otherwise it is not possible to fix building an OTA package without a
recovery image. Due to this, a special system to automatically patch
repositories from the full tree when needed before building was
introduced. It is an ugly hack, but it is a necessary hack :(