Enable loading of libGLES_android.so when using software OpenGL ES.
Before Android 4.4 the default software OpenGL ES library from Android
was loaded when no other OpenGL ES library was found. In 4.4, however,
it is always ignored on a real device (it is taken into account, though,
when the emulator is being used), which causes the graphics system to
fail if there is no other OpenGL ES library available. Therefore, the
EGL loader must be patched to take libGLES_android.so into account when
the device declares that it uses the software OpenGL ES library.