Add FFmpeg codecs to Stagefright configuration.
CyanogenMod 11.0 added several FFmpeg codecs that are automatically
built when inheriting from the common CyanogenMod configuration
(vendor/cm/config/common.mk). However, they must be explicitly added to
the Stagefright configuration in order to be able to use them.
Therefore, the default Stagefright configuration from goldfish was
extended with the entries for the FFmpeg codecs included in the
Stagefright configuration from mako.