Add parameter to abort build when build system was patched.
When patching files belonging to the build system, the build must be
stopped and then launched again. Otherwise, the changes applied to the
build system may not be taken into account. To do this, the functions
that perform the patching, "patch-repository" and
"reverse-patch-repository", now can be called with an optional third
parameter to know whether the patch was applied or not. That parameter
is later checked and, if any of the patches set it, then the build is
aborted (so the parameter should be used only with those patches that
modify the build system).