Do not create backups of patched files.
When a patch is applied successfully but within certain offset from the
lines stated in the patch a backup of the original file was created. The
backups are not needed and they just add noise to the repository, as the
original files come from a version control system, so now the backups
are no longer created.