Add patch to use RGB 565 in Stagefight software renderer.
By default, the built-in software video codecs from Stagefright use the
YUV420/YV12 pixel format. That format is not supported by the gralloc
module nor the software OpenGL ES library, so the videos could not be
viewed using the components provided by this FOSS device tree. However,
instead of having to add support for that format to the gralloc module
and the software OpenGL ES library, the Stagefright software renderer
can be patched to perform the same conversion to RGB 565 used in the
emulator also when the software OpenGL ES library is being used.