Add support for MultiROM.
This commit provides the device files needed to build a Team Win
Recovery Project recovery with the MultiROM extension, plus the MultiROM
bootloader, its installer and its uninstaller, all of them on OmniROM
4.4.

Currently the versions included in the official MultiROM repositories
are not compatible with the Fairphone 1; the branches "master-fp1" of
libbootimg and multirom, and the branch "android-6.0-fp1" of
Team-Win-Recovery-Project, all from https://gitlab.com/danxuliu, must be
used instead. The needed changes were sent upstream, so in the future it
might be possible to build them using the official MultiROM repositories
too.